Rename internal/users.py to internal/auth.py
