+++ user-api/container/user_api/routers/main.py
@@ -7,7 +7,7 @@ from psycopg.types.json import set_json_dumps, set_json_loads
 import orjson
 
 from user_api import config
-from user_api.internal.users import clean_sessions_loop
+from user_api.internal.auth import clean_sessions_loop
 from user_api.routers import users
